Refactoring style's pathes
Place components' style files in the same directory of the component to keep it simple.
