+++ src/components/commitRow/commit_row.scss
