+++ src/components/loading/loading.scss
